requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
django-extensions==3.2.1
ipython==8.7.0
django-colorfield==0.8.0